01-MODELS/2-chat-models/3-gemini.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model = ChatGoogleGenerativeAI(model='gemini-2.0-flash-lite')

result = model.invoke('what is a pizza?')

print(result.content)